refactor: 쿼리 훅 api 폴더 하위로 이동
